fix(web): eliminate all State & Effects errors
- visual-timer: merge the two timer-finished effects into a single
  useEffectEvent handler driven by the completion event (clears the
  adjust-state-on-prop-change and missing-cleanup findings)
- inline-quiz: hydrate sessionStorage progress via lazy initial state,
  remount per step with a key instead of a hydrate effect
- badge-celebration: play the burst on mount, remount via key to replay
- family-invite/share dialogs: extract bodies mounted only while open so
  their state initializes fresh, removing the reset effects
- delete unused BarkleyQuizDialog (superseded by InlineQuiz)
